Meta clean up
Cleaned the files in the project folder
