--- src/is2lab1/Rent.java
@@ -1,80 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package is2lab1;
-
-import java.time.LocalDate;
-
-/**
- *
- * @author Lokem
- */
-public class Rent {
-    int rentID;
-    int clientID;
-    int objectID;
-    LocalDate rentStart;
-    LocalDate rentEnd;
-    double startUpPrice;
-    double totalPrice;
-    
-    public double getStartUpPrice() {
-        return startUpPrice;
-    }
-
-    public void setStartUpPrice(double startUpPrice) {
-        this.startUpPrice = startUpPrice;
-    }
-
-    public double getTotalPrice() {
-        return totalPrice;
-    }
-
-    public void setTotalPrice(double totalPrice) {
-        this.totalPrice = totalPrice;
-    }
-    
-    public int getRentID() {
-        return rentID;
-    }
-
-    public void setRentID(int rentID) {
-        this.rentID = rentID;
-    }
-
-    public int getClientID() {
-        return clientID;
-    }
-
-    public void setClientID(int clientID) {
-        this.clientID = clientID;
-    }
-
-    public int getObjectID() {
-        return objectID;
-    }
-
-    public void setObjectID(int objectID) {
-        this.objectID = objectID;
-    }
-
-    public LocalDate getRentStart() {
-        return rentStart;
-    }
-
-    public void setRentStart(LocalDate rentStart) {
-        this.rentStart = rentStart;
-    }
-
-    public LocalDate getRentEnd() {
-        return rentEnd;
-    }
-
-    public void setRentEnd(LocalDate rentEnd) {
-        this.rentEnd = rentEnd;
-    }
-    
-    
-}
